[INJIMOB-3181]: refactor deep link flow for future features
